Modernize RakSAMP client and server

+++ client/src/math_stuff.h
@@ -48,4 +48,4 @@ void MakeOrthMatrix(float fInW, float fInX, float fInY, float fInZ,
 					float *m10, float *m11, float *m12,
 					float *m20, float *m21, float *m22);
 
-void BIG_NUM_MUL(unsigned long in[5], unsigned long out[6], unsigned long factor);
+void BIG_NUM_MUL(unsigned int in[5], unsigned int out[6], unsigned int factor);
